Added test for parts of IF_DEF

+++ test_ccdb_dump.py
@@ -24,7 +24,8 @@ class mkdtemp(object):
 
 
     def __exit__(self, type, value, traceback):
-        shutil.rmtree(self.dirpath)
+        if type is None:
+            shutil.rmtree(self.dirpath)
